backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
openpyxl==3.1.5
aiofiles==24.1.0
pydantic==2.9.0
